pyproject.toml
[tool.poetry]
name = "hmppctshmppc-backend"
version = "0.1.0"
description = "HMPPCTSHMPPC Backend"
authors = ["Prasanta Kakati <prasantakakati1994@gmail.com>"]
license = "Not open source"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.2"
uwsgi = "^2.0.18"
djangorestframework = "^3.12.4"
requests = "^2.25.1"
celery = "^5.0.5"
django-cors-headers = "^3.7.0"
redis = "^3.5.3"
sentry-sdk = "^1.3.1"

[tool.poetry.dev-dependencies]
pep8 = "^1.7.1"
autopep8 = "^1.5"
ipython = "^7.11.1"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
